@@ -3,22 +3,18 @@ description: Demonstrates how to use the flutter_sms plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0-259.8.beta <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
-
 dev_dependencies:
+  flutter_sms:
+    path: ../
   flutter_test:
     sdk: flutter
 
-  flutter_sms:
-    path: ../
-
 flutter:
-
   uses-material-design: true
 
