@@ -1,23 +1,22 @@
 name: flutter_sms
 description: A Flutter plugin to Send SMS and MMS on iOS and Android. If iMessage is enabled it will send as iMessage on iOS. This plugin must be tested on a real device on iOS.
-version: 2.2.0+1
+version: 2.3.1
 author: Rody Davis <rody.davis.jr@gmail.com>
 homepage: https://github.com/rodydavis/plugins
 repository: https://github.com/fluttercommunity/flutter_sms
 maintainer: Rody Davis (@rodydavis)
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.12.8 <2.0.0"
+  sdk: ">=2.12.0-259.8.beta <3.0.0"
+  flutter: ^1.10.0
 
 dependencies:
   flutter:
     sdk: flutter
-  url_launcher: ^6.0.2
   flutter_web_plugins:
     sdk: flutter
   plugin_platform_interface: ^2.0.0
-  flutter_user_agent: ^1.2.2
+  url_launcher: ^6.0.3
 
 dev_dependencies:
   flutter_test:
@@ -31,8 +30,6 @@ flutter:
         pluginClass: FlutterSmsPlugin
       ios:
         pluginClass: FlutterSmsPlugin
-      # macos:
-      #   pluginClass: FlutterSmsPlugin
-      # web:
-      # pluginClass: FlutterSmsPlugin
-      # fileName: flutter_sms_web.dart
+      web:
+        pluginClass: FlutterSmsPlugin
+        fileName: flutter_sms_web.dart
