@@ -1,11 +1,12 @@
 import 'dart:async';
-import 'dart:io';
+import 'dart:io' show Platform;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_user_agent/flutter_user_agent.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 import 'package:url_launcher/url_launcher.dart';
+
+import 'user_agent/io.dart' if (dart.library.html) 'user_agent/web.dart';
 
 const MethodChannel _channel = MethodChannel('flutter_sms');
 
@@ -32,69 +33,52 @@ class FlutterSmsPlatform extends PlatformInterface {
   }
 
   Future<String> sendSMS({
-    @required String message,
-    @required List<String> recipients,
+    required String message,
+    required List<String> recipients,
   }) {
-    var mapData = Map<dynamic, dynamic>();
-    mapData["message"] = message;
+    final mapData = <dynamic, dynamic>{};
+    mapData['message'] = message;
     if (!kIsWeb && Platform.isIOS) {
-      mapData["recipients"] = recipients;
-      return _channel.invokeMethod<String>('sendSMS', mapData);
+      mapData['recipients'] = recipients;
+      return _channel
+          .invokeMethod<String>('sendSMS', mapData)
+          .then((value) => value ?? 'Error sending sms');
     } else {
-      String _phones = recipients.join(";");
-      mapData["recipients"] = _phones;
-      return _channel.invokeMethod<String>('sendSMS', mapData);
+      String _phones = recipients.join(';');
+      mapData['recipients'] = _phones;
+      return _channel
+          .invokeMethod<String>('sendSMS', mapData)
+          .then((value) => value ?? 'Error sending sms');
     }
   }
 
   Future<bool> canSendSMS() {
-    return _channel.invokeMethod<bool>('canSendSMS');
+    return _channel
+        .invokeMethod<bool>('canSendSMS')
+        .then((value) => value ?? false);
   }
 
-  Future<bool> launchSmsMulti(List<String> numbers, [String body]) {
-    if (numbers == null || numbers.length == 1) {
-      return launchSms(numbers?.first, body);
+  Future<bool> launchSmsMulti(List<String> numbers, [String? body]) {
+    if (numbers.length == 1) {
+      return launchSms(numbers.first, body);
     }
-    String _phones = numbers.join(";");
+    String _phones = numbers.join(';');
     if (body != null) {
       final _body = Uri.encodeComponent(body);
-      return launch('sms:/open?addresses=$_phones${seperator}body=$_body');
+      return launch('sms:/open?addresses=$_phones${separator}body=$_body');
     }
     return launch('sms:/open?addresses=$_phones');
   }
 
-  Future<bool> launchSms(String number, [String body]) {
-    if (number == null) {
-      number = '';
-    }
+  Future<bool> launchSms(String? number, [String? body]) {
+    // ignore: parameter_assignments
+    number ??= '';
     if (body != null) {
       final _body = Uri.encodeComponent(body);
-      return launch('sms:/$number${seperator}body=$_body');
+      return launch('sms:/$number${separator}body=$_body');
     }
     return launch('sms:/$number');
   }
 
-  String get seperator => isCupertino() ? '&' : '?';
-
-  bool isCupertino() {
-    if (kIsWeb) {
-      final _devices = [
-        'iPad Simulator',
-        'iPhone Simulator',
-        'iPod Simulator',
-        'iPad',
-        'iPhone',
-        'iPod',
-        'Mac OS X',
-      ];
-      final String _agent = FlutterUserAgent.webViewUserAgent;
-      for (final device in _devices) {
-        if (_agent.contains(device)) {
-          return true;
-        }
-      }
-      return false;
-    }
-    return Platform.isIOS || Platform.isMacOS;
-  }
+  String get separator => isCupertino() ? '&' : '?';
 }
